requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
jinja2==3.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.1.1
requests==2.32.5

# 测试依赖
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0